% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_file_sig.R
\name{add_file_sig}
\alias{add_file_sig}
\title{Add a file signature to an existing registry entry}
\usage{
add_file_sig(
  registry_entry,
  id,
  extension,
  filename_pattern = NULL,
  date = NULL
)
}
\arguments{
\item{registry_entry}{\link{hs_registry_entry} object created with
\code{\link[=new_registry_entry]{new_registry_entry()}}.}

\item{id}{character vector of length one. Identifier of embedded metadata signature.
Must use snake_case (lowercase letters, numbers, and underscores only)}

\item{extension}{character vector of length one. Expected file extension
(e.g. "CR2", "NEF", "DNG").}

\item{filename_pattern}{optional character vector of length one. Usual
filename structure \emph{excluding the file extension}, typically expressed as a
regular expression describing the default camera-generated basename. (e.g.
\code{"^IMG_[0-9]{4}$"} matches basenames such as \code{IMG_0001},
\code{IMG_0234}).}

\item{date}{optional \code{\link{Date}} object. Date when the specification was added
to the registry. If \code{NULL}, the current system date is used automatically.}
}
\value{
A \link{hs_registry_entry} object with the \link{file_sig} appended.
}
\description{
Add a \link{file_sig} to a \link{hs_registry_entry} object.
}
\examples{
foo <- new_registry_entry(
  "Nikon_Coolpix5700.FCE9.CIEFAP",
  body = "E5700",
  body_manufacturer = "NIKON CORP",
  body_serial = "7053067",
  lens = "Zoom Nikkor ED 8.9-71.2mm 1:2.8-4.2",
  lens_manufacturer = "NIKON CORP",
  auxiliary_lens = "Fisheye Converter FC-E9 0.2x",
  auxiliary_lens_manufacturer = "NIKON CORP",
  institution = "CIEFAP"
)

foo <- add_file_sig(
  foo,
  id = "raw",
  extension = "NEF",
  filename_pattern = "^DSCN[0-9]{4}$"
)

foo
summary(foo)
}
