% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_doc_Canon_EOS_5D-Sigma_8mm-EMY.R
\docType{data}
\name{Canon_EOS_5D.Sigma_8mm.EMY}
\alias{Canon_EOS_5D.Sigma_8mm.EMY}
\title{Specs for}
\format{
An object of class \code{hemispherical_camera_registry} containing
one geometry specification and one associated radiometry specification.
}
\usage{
Canon_EOS_5D.Sigma_8mm.EMY
}
\description{
A complete set of geometric and radiometric specifications for
}
\details{
\insertCite{Diaz2024;textual}{rcaiman}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
