% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class_documentation.R
\docType{class}
\name{file_sig}
\alias{file_sig}
\title{File signature}
\description{
Represent a file signature used by \pkg{rcaiman} to facilitate
file retrieval tasks.
}
\note{
Can be used independently to supply arguments to \pkg{rcaiman} core
functions. However, its domain of validity remains tied as declared in
\emph{Domain of application}.
}
\section{Definition}{

A \code{file_sig} describes image files as observed externally (e.g. filename
structure, extension).
}

\section{Domain of application}{

A file signature acquires declared validity only within a
\link{hs_registry_entry}.
}

\section{Structure}{

A \code{file_sig} is an S3 object with the following fields:

\describe{
\item{id}{Identifier (snake_case).}
\item{extension}{Expected file extension (e.g., \code{"NEF"}, \code{"CR2"}).}
\item{filename_pattern}{Optional regular expression describing expected basename structure.}
\item{date}{Date of signature declaration.}
}
}

\section{Generation}{

A file signature is created via \code{\link[=add_file_sig]{add_file_sig()}}. Registration is declarative
and does not perform file matching.
}

\section{Retrieving}{

A \code{file_sig} may be retrieved using \code{\link[=get_file_sig]{get_file_sig()}}.
}

