% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcaiman.registry-package.R
\docType{package}
\name{rcaiman.registry-package}
\alias{rcaiman.registry}
\alias{rcaiman.registry-package}
\title{rcaiman.registry: Instrument Registry for CAnopy IMage ANalysis}
\description{
Provides a structured registry of instrument specifications and calibration metadata for canopy image analysis within the rcaiman ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/GastonMauroDiaz/rcaiman.registry/issues}
}

}
\author{
\strong{Maintainer}: Gastón Mauro Díaz \email{gastonmaurodiaz@gmail.com} (\href{https://orcid.org/0000-0002-0362-8616}{ORCID})

}
\keyword{internal}
