% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_registry_entry.R
\name{export_registry_snapshot}
\alias{export_registry_snapshot}
\alias{read_registry_snapshot}
\title{Export and read registry snapshots}
\usage{
export_registry_snapshot(
  registry_entry,
  embedded_metadata_sig,
  geometry_spec,
  interpretive_constraint,
  flat_field_correction,
  file
)

read_registry_snapshot(file)
}
\arguments{
\item{registry_entry}{\link{hs_registry_entry} object created with
\code{\link[=new_registry_entry]{new_registry_entry()}}.}

\item{embedded_metadata_sig}{character vector of length one. Identifier of
embedded metadata signature.}

\item{geometry_spec}{Character vector of length one.
Identifier of geometry specification.}

\item{interpretive_constraint}{optional character vector of length one.
Identifier of a \link{radiometry_spec} of type \code{interpretive_constraint}.
Provide \code{NULL} in the case of JPEG files or when the use of RAW metadata
via \emph{rawpy} is preferred.}

\item{flat_field_correction}{character vector of length one.
Identifier of a \link{radiometry_spec} of type \code{flat_field_correction}.}

\item{file}{character vector of length one. Path to the YAML file used to
write or read the snapshot.}
}
\value{
\code{export_registry_snapshot()} no return value. Called for side effects.

\code{read_registry_snapshot()} returns an object of class \code{\link{hs_registry_entry}}.
}
\description{
Create and read portable snapshots of registry entries stored as YAML files.
}
\details{
A registry snapshot is a YAML representation of a specific subset of
an \link{hs_registry_entry} object. It captures a single combination of
\link{embedded_metadata_sig}, \link{geometry_spec}, and \link{radiometry_spec} with
\verb{type = "interpretive_constraint} and with \code{type = "flat_field_correction"},
together with intrument metadata and minimal metadata about the
exporting environment.

This mechanism provides a lightweight way to declare image geometric
projection and radiometric correction without requiring the full registry
infrastructure. In this way, the YAML file can act as the metadata of
pre-processed images.

Moreover, \code{\link[=read_registry_snapshot]{read_registry_snapshot()}} reconstructs an \link{hs_registry_entry}
object from a snapshot YAML file, making this format also useful to store and
distribute single-branch registry estries.
}
\examples{

# Build the object to export ----------------------------------------------

foo <- new_registry_entry(
  "Nikon_Coolpix5700.FCE9.CIEFAP",
  body = "E5700",
  body_manufacturer = "NIKON CORP",
  body_serial = "7053067",
  lens = "Zoom Nikkor ED 8.9-71.2mm 1:2.8-4.2",
  lens_manufacturer = "NIKON CORP",
  auxiliary_lens = "Fisheye Converter FC-E9 0.2x",
  auxiliary_lens_manufacturer = "NIKON CORP",
  institution = "CIEFAP"
)

foo <- add_file_sig(
  foo,
  id = "raw",
  extension = "NEF",
  filename_pattern = "^DSCN[0-9]{4}$"
)

foo <- add_embedded_metadata_sig(
  foo,
  id = "exif_01",
  namespace = "exif",
  dim = c(1288, 962),
  validated_with_rawpy = "0.19.0",
  validated_with_libraw = "0.21.1",
  rules = list(
    "Camera Model Name" = "E5700",
    "Software" = "E5700v1.1",
    "CFA Pattern" = "[Yellow,Cyan][Green,Magenta]",
    "Compression" = "Uncompressed",
    "Bits Per Sample" = "12",
    "Quality" = "RAW"
  ),
  file_sig = "raw"
)

foo <- add_geometry_spec(
  foo,
  embedded_metadata_sig = "exif_01",
  id = "simple_method",
  model = "radial_projection",
  parameters = c(0.6380,  0.0307, -0.0200),
  zenith_col_row = c(645, 494),
  horizon_radius = 378,
  is_horizon_circle_clipped = FALSE,
  max_zenith_angle = 94.8,
  notes = "Calibration documented in doi:10.1016/j.agrformet.2024.110020",
  contact_information = "gastonmaurodiaz@gmail.com"
)

foo <- add_radiometry_spec(
  foo,
  embedded_metadata_sig = "exif_01",
  geometry_spec = "simple_method",
  id = "spectral_bands",
  type = "interpretive_constraint",
  cfa_pattern = matrix(c("Green", "Yellow",
                         "Magenta", "Cyan"), byrow = TRUE, ncol = 2),
  spectral_mapping = list(Red = "(Yellow + Magenta)/2",
                          Green = "Green",
                          Blue = "Cyan"),
  offset_value = list("100" = 0),
  contact_information = "gastonmaurodiaz@gmail.com"
)

foo <- add_radiometry_spec(
  foo,
  embedded_metadata_sig = "exif_01",
  geometry_spec = "simple_method",
  id = "simple_method",
  type = "flat_field_correction",
  model = "flat_field_simple_polynomial",
  parameters = list("5.0" = c(0.0638, -0.101)),
  notes = "Calibration documented in doi:10.1016/j.agrformet.2024.110020",
  contact_information = "gastonmaurodiaz@gmail.com"
)

export_registry_snapshot(foo,
                         "exif_01",
                         "simple_method",
                         "spectral_bands",
                         "simple_method",
                         "registry_snapshot.yaml")

read_registry_snapshot(system.file("external/registry_snapshot.yaml", package = "rcaiman.registry"))

}
\seealso{
\code{\link[=add_embedded_metadata_sig]{add_embedded_metadata_sig()}}, \code{\link[=add_geometry_spec]{add_geometry_spec()}},
\code{\link[=add_radiometry_spec]{add_radiometry_spec()}}
}
