% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry_spec.R
\name{add_geometry_spec}
\alias{add_geometry_spec}
\title{Add a geometry spec to an existing registry}
\usage{
add_geometry_spec(
  registry,
  id,
  date = NULL,
  lens_coef,
  zenith_colrow,
  horizon_radius,
  is_horizon_circle_clipped,
  max_zenith_angle = NULL,
  dim = NULL,
  firmware_version = NULL,
  notes = NULL,
  contact_information = NULL
)
}
\arguments{
\item{registry}{\code{hemispherical_camera_registry} object created with
\code{\link[=new_registry]{new_registry()}}.}

\item{id}{character vector of length one. Identifier of the geometry spec.
Must use snake_case (lowercase letters, numbers, and underscores only)}

\item{date}{optional \code{\link{Date}} object. Date when the specification was added
to the registry. If \code{NULL}, the current system date is used automatically.}

\item{lens_coef}{numeric vector. Lens projection coefficients describing the
relationship between image radius and zenith angle. These coefficients
typically originate from \code{\link[rcaiman:calibrate_lens]{rcaiman::calibrate_lens()}} or compatible
procedures following the same conventions. The coefficients must pass
\code{\link[rcaiman:test_lens_coef]{rcaiman::test_lens_coef()}}.}

\item{zenith_colrow}{numeric vector of length 2. Column and row coordinates
of the optical center (zenith) in pixel units following the conventions
used in \code{rcaiman}.}

\item{horizon_radius}{integer-like numeric vector of length one. Radius of the
hemispherical image in pixel units, measured from the optical center
(\code{zenith_colrow}) to the image location corresponding to a zenith angle of
90 deg (i.e. the horizon), and following the conventions used in \code{rcaiman}
(i.e., \code{2 * horizon_radius} must be an even integer, see
\link[rcaiman:zenith_image]{rcaiman::zenith_image}).}

\item{is_horizon_circle_clipped}{logical vector of length one. Indicates
whether the horizon circle (90° zenith angle), obtained via model
extrapolation when necesary, is clipped by the image matrix. If \code{FALSE},
the geometry specification asserts that the complete horizon circle is
fully included within the image extent, and this condition may be validated
geometrically. If \code{TRUE}, horizon clipping is expected and no such
validation is performed.}

\item{max_zenith_angle}{optional numeric vector of length one. Maximum zenith
angle (deg) that the system is capable of observing.}

\item{dim}{optional integer-like numeric vector of length two. Width and height in
pixels of the raster from which \code{zenith_colrow} was derived.}

\item{firmware_version}{optional character vector of length one.
Firmware version of the image acquisition system at the time the
calibration images were acquired. This information is stored for
documentation and traceability purposes.}

\item{notes}{optional character vector of length one. Free-form notes
providing context, scope, assumptions, or references relevant to this
specification (e.g. publications).}

\item{contact_information}{optional character vector of length one. Contact
information for the specification maintainer or responsible person.}
}
\value{
A \code{hemispherical_camera_registry} object with the geometry specification
appended.
}
\description{
Add a geometry specification to an existing hemispherical
camera registry.
}
\examples{
foo <- new_registry(
  "Nikon_D610.Nikkor_8mm.CIEFAP",
  body = "D610",
  body_manufacturer = "NIKON CORP",
  body_serial = "9023728",
  lens = "AF-S FISHEYE NIKKOR 8-15mm 1:3.5-4.5E ED",
  lens_manufacturer = "NIKON CORP",
  lens_serial = "210020",
  institution = "CIEFAP"
)

foo <- add_geometry_spec(
  foo,
  id = "simple_method",
  lens_coef = signif(c(1306,24.8,-56.2)/1894,3),
  zenith_colrow = c(1500, 997),
  horizon_radius = 947,
  is_horizon_circle_clipped = FALSE,
  max_zenith_angle = 92.8,
  dim = c(3040, 2014),
  firmware_version = "1.01",
  notes = "Calibration documented in doi:10.1016/j.agrformet.2024.110020",
  contact_information = "gastonmaurodiaz@gmail.com"
)
}
