% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry_spec.R
\name{add_geometry_spec}
\alias{add_geometry_spec}
\title{Add a geometry spec to an existing registry}
\usage{
add_geometry_spec(
  registry,
  id,
  date,
  lens_coef,
  zenith_colrow,
  horizon_radius,
  max_zenith_angle = NULL,
  dim = NULL,
  notes = NULL,
  contact_information = NULL
)
}
\arguments{
\item{registry}{\code{hemispherical_camera_registry} object created with
\code{\link[=new_registry]{new_registry()}}.}

\item{id}{character vector of length one. Identifier of the geometry spec.}

\item{date}{a \code{\link{Date}} object indicating when the calibration was performed.}

\item{lens_coef}{numeric vector. Lens projection coefficients describing the
relationship between image radius and zenith angle. These coefficients
typically originate from \code{\link[rcaiman:calibrate_lens]{rcaiman::calibrate_lens()}} or compatible
procedures following the same conventions. The coefficients must pass
\code{\link[rcaiman:test_lens_coef]{rcaiman::test_lens_coef()}}.}

\item{zenith_colrow}{numeric vector of length 2. Column and row coordinates
of the optical center (zenith) in pixel units following the conventions
used in \code{rcaiman}.}

\item{horizon_radius}{integer-like numeric vector of length one. Radius of the
hemispherical image in pixel units, measured from the optical center
(\code{zenith_colrow}) to the image location corresponding to a zenith angle of
90 deg (i.e. the horizon), and following the conventions used in \code{rcaiman}
(i.e., \code{2 * horizon_radius} must be an even integer, see
\link[rcaiman:zenith_image]{rcaiman::zenith_image}).}

\item{max_zenith_angle}{optional numeric vector of length one. Maximum zenith
angle (deg) that the system is capable of observing.}

\item{dim}{integer-like numeric vector of length two. Width and height in
pixels of the raster from which \code{zenith_colrow} was derived.}

\item{notes}{optional character vector of length one. Free-form notes
describing the calibration procedure, assumptions, or references (e.g.
publications).}

\item{contact_information}{optional character vector of length one. Contact
details of the author of the calibration.}
}
\value{
A \code{hemispherical_camera_registry} object with the geometry specification
appended.
}
\description{
Add a geometric calibration specification to an existing hemispherical
camera registry.
}
