% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_radiometry_spec.R
\name{add_radiometry_spec}
\alias{add_radiometry_spec}
\title{Add a calibration spec to an existing geometry spec}
\usage{
add_radiometry_spec(
  registry,
  geometry_id,
  id,
  type = "vignetting",
  method,
  model,
  parameters,
  date,
  notes = NULL,
  contact_information = NULL
)
}
\arguments{
\item{registry}{\code{hemispherical_camera_registry} object created with
\code{\link[=new_registry]{new_registry()}} and containing the geometry specification identified by
\code{geometry_id}.}

\item{geometry_id}{character vector of length one. Identifier of the geometry
specification to which the radiometry spec will be attached.}

\item{id}{character vector of length one. Identifier of the radiometry spec.}

\item{type}{character vector of length one. Type of radiometric correction.
Only "vignetting" is currently supported.}

\item{method}{character vector of length one. Method used to acquire the
primary data needed for radiometric calibration. Supported methods are
"simple" and "photometric_sphere".}

\item{model}{character vector of length one. Model used to develop a
radiometric correction. Only "polynomial" is currently supported via
\link[rcaiman:correct_vignetting]{rcaiman::correct_vignetting}.}

\item{parameters}{numeric vector. Model parameters defining the radiometric
correction, following the conventions of the selected \code{model}.}

\item{date}{a \code{\link{Date}} object indicating when the calibration was performed.}

\item{notes}{optional character vector of length one. Free-form notes
describing the calibration procedure, assumptions, or references (e.g.
publications).}

\item{contact_information}{optional character vector of length one. Contact
details of the author of the calibration.}
}
\value{
A \code{hemispherical_camera_registry} object with the radiometry specification
appended to the selected geometry specification.
}
\description{
Add a radiometric calibration specification to an existing geometry
specification in a hemispherical camera registry.
}
