% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class_documentation.R
\docType{class}
\name{embedded_metadata_sig}
\alias{embedded_metadata_sig}
\title{Embedded metadata signature}
\description{
Represent an embedded metadata signature used by \pkg{rcaiman} to
identify the instruments that captured an image and their configuration
state at acquisition time.
}
\note{
Can be used independently to supply arguments to \pkg{rcaiman} core
functions. However, its domain of validity remains tied as declared in
\emph{Domain of application}.
}
\section{Definition}{

An \code{embedded_metadata_sig} represents a configuration state of an
instrument as expressed through embedded metadata.
}

\section{Domain of application}{

An embedded metadata signature acquires full meaning only within a
\link{hs_registry_entry} via \code{intrument_metada}, which purpouse is to
offers keys to identify the p physical acquisition system that can produce
such data.
}

\section{Structure}{


An \code{embedded_metadata_sig} is an S3 object with the following fields:

\describe{
\item{id}{Identifier (snake_case).}
\item{namespace}{Lowercase identifier of metadata namespace (e.g., \code{"exif"}).}
\item{dim}{Integer-like vector of length two specifying image width and height.}
\item{rules}{Named list of expected tag–value pairs.}
\item{date}{Date of signature declaration.}
\item{file_sig}{Optional foreign key referencing a \link{file_sig} of the \link{hs_registry_entry} object.}
\item{geometry}{Named list of \link{geometry_spec} objects.}
}
}

\section{Generation and validation}{


An embedded metadata signature is created via \code{\link[=add_embedded_metadata_sig]{add_embedded_metadata_sig()}}.

Registration validates structural consistency (identifier format, rule
structure, optional file signature linkage), but does not perform metadata
matching at creation time.
}

\section{Retrieving}{

A \code{embedded_metadata_sig} may be retrieved using \code{\link[=get_embedded_metadata_sig]{get_embedded_metadata_sig()}}.
}

