% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_embedded_metadata_sig.R
\name{get_embedded_metadata_sig}
\alias{get_embedded_metadata_sig}
\title{Extract embedded metadata signature}
\usage{
get_embedded_metadata_sig(registry_entry, embedded_metadata_sig)
}
\arguments{
\item{registry_entry}{\link{hs_registry_entry} object created with
\code{\link[=new_registry_entry]{new_registry_entry()}}.}

\item{embedded_metadata_sig}{character vector of length one. Identifier of
embedded metadata signature.}
}
\description{
Retrieve a embedded metadata signature within a registry entry.
}
\examples{

# Build de object ---------------------------------------------------------

foo <- new_registry_entry(
  "Nikon_Coolpix5700.FCE9.CIEFAP",
  body = "E5700",
  body_manufacturer = "NIKON CORP",
  body_serial = "7053067",
  lens = "Zoom Nikkor ED 8.9-71.2mm 1:2.8-4.2",
  lens_manufacturer = "NIKON CORP",
  auxiliary_lens = "Fisheye Converter FC-E9 0.2x",
  auxiliary_lens_manufacturer = "NIKON CORP",
  institution = "CIEFAP"
)

foo <- add_embedded_metadata_sig(
  foo,
  id = "exif_01",
  namespace = "exif",
  dim = c(1288, 962),
  rules = list(
    "Camera Model Name" = "E5700",
    "Software" = "E5700v1.1",
    "CFA Pattern" = "[Yellow,Cyan][Green,Magenta]",
    "Compression" = "Uncompressed",
    "Bits Per Sample" = "12",
    "Quality" = "RAW"
  )
)

foo <- add_geometry_spec(
  foo,
  embedded_metadata_sig = "exif_01",
  id = "simple_method",
  model = "radial_projection",
  parameters = c(0.6380,  0.0307, -0.0200),
  zenith_col_row = c(645, 494),
  horizon_radius = 378,
  is_horizon_circle_clipped = FALSE,
  max_zenith_angle = 94.8,
  notes = "Calibration documented in doi:10.1016/j.agrformet.2024.110020",
  contact_information = "gastonmaurodiaz@gmail.com"
)

foo <- add_radiometry_spec(
  foo,
  embedded_metadata_sig = "exif_01",
  geometry_spec = "simple_method",
  id = "spectral_bands",
  type = "interpretive_constraint",
  cfa_pattern = matrix(c("Green", "Yellow",
                         "Magenta", "Cyan"), byrow = TRUE, ncol = 2),
  spectral_mapping = list(Red = "(Yellow + Magenta)/2",
                          Green = "Green",
                          Blue = "Cyan"),
  offset_value = list("100" = 0),
  contact_information = "gastonmaurodiaz@gmail.com"
)

# -------------------------------------------------------------------------

em <- get_embedded_metadata_sig(foo, "exif_01")
}
